requirements.txt
yandexcloud==0.194.0
python-telegram-bot==20.0